Changed fetch_data.py to main.py
Realized there were 2 api so i deleted the old main file that used
fastapi and instead renamed fetch_data.py to main.py
In main.py renamed news_summary route to search route,
added summarize route endpoint that connects to GeminiSumarize.
Added swagger to main.py and docstrings to both routes so that they are
visible in swagger.
