+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 venv/
-
+database/
+__pycache__/
